// src/app/dashboard/objects/enums/enums.enum.ts
export enum VotingTypes {
  LIKE = '1',
  DISLIKE = '2'
}

export enum EmoticonTypes {
  EXCITED = '1',
  UPSIDE_DOWN = '2',
  THINKING = '3',
  FRUSTRATED = '4'
}

export enum MessageType {
  MESSAGE,
  ERROR
}

export enum DocumentType {
  ANNOUNCEMENT = 'announcement',
  QUESTION = 'question',
  QUICK_ANNOUNCEMENT = 'quick-announcement'
}

export enum Privileges {
  CAN_EDIT = 'CAN_EDIT',
  CAN_PUBLISH = 'CAN_PUBLISH'
}

export enum SortTypeList {
  DATE_DESCENDING = 'BY_DATE_DESCENDING',
  TOP = 'TOP'
}

export enum SortTypeListDesc {
  DATE_DESCENDING = 'najnovšie príspevky',
  TOP = 'top príspevky'
}