// src/app/shared/objects/enums/recipients.enum.ts
export enum RecipientsCollectionType {
  TEAMS = 'TEAMS',
  FUNCTIONS = 'FUNCTIONS',
  UNITS = 'UNITS',
  GROUPS = 'GROUPS',
  USERS = 'USERS'
}

export enum RecipientsCollectionTypeDesc {
  TEAMS = 'Tímy',
  FUNCTIONS = 'Funkcie',
  UNITS = 'Útvary',
  GROUPS = 'Skupiny',
  USERS = 'Osoby'
}